requirements.txt
torch>=2.1.0
transformers>=4.45.0
datasets>=2.18.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
peft>=0.12.0
trl>=0.9.0
ragas>=0.1.10
langchain-openai>=0.1.0
rouge-score>=0.1.2
nltk>=3.8.0
pyyaml>=6.0
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=4.25.0
matplotlib>=3.8.0
huggingface-hub>=0.23.0
